kNN method to compare

+++ src/main/java/pl/edu/ug/FullProbRule.java
@@ -10,7 +10,7 @@ public class FullProbRule extends Rule {
 
         if (img[row][col] != 0) return img[row][col];
 
-        Map<Byte, Byte> countedClasses = countClasses(img, row, col);
+        Map<Byte, Integer> countedClasses = countClasses(img, row, col);
 
 
         int all = countedClasses.get((byte) 0) + countedClasses.get((byte) 1) + countedClasses.get((byte) 2);
